+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.6.0] - 2018-12-19
 ### Added
 - Added ability to override the default password length
 - Merge the pbs-ruby gem removing that as a dependency, but adding FFI
@@ -138,7 +139,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release!
 
-[Unreleased]: https://github.com/OSC/ood_core/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/OSC/ood_core/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/OSC/ood_core/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/OSC/ood_core/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/OSC/ood_core/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/OSC/ood_core/compare/v0.3.0...v0.4.0